Add ApplyOctomapMeshMask Action Server
This implementation takes the octomap as a point cloud that is currently being published. If it is not found in 5 seconds, it returns failed. It also takes a path to PLY mesh file and returns a path to PLY mesh file. Modifying it to return a Tesseract_geometry::Mesh would not be hard.