Add PointCloudXYZItoXYZRGB
Some cameras like the realsense publish pointclouds as XYZI when colorizing them with greyscale. Since all of this code expects XYZRGB, I made this to convert it. Eventually, the conversion logic may be put directly in the image extractor node to reduce overhead, but this is convenient for debugging at the moment.