Turn on dynamic memory allocation on the GPU
This allows GPU memory to dynamically grow. This is a workaround to fix this issue on RTX cards. https://github.com/tensorflow/tensorflow/issues/24496. However, this can be problematic when sharing memory between applications. TODO: Check and see if issue 24496 has been closed, and change this. Note that since Tensorflow 1.15 is the final 1.x release, this might never happen until this code is upgraded to tensorflow 2.x